dark mode fixes, added ascent color picker with logic
